Change some whole number types and casts
Replace Cardinal & other types with Int32:

  Every use of Cardinal related to the C size_t, which in turn related
  to file sizes and positions or internal data structures that depend on
  file sizes. But BDiff can't handle files >= 2GiB, i.e. files must be
  less than MaxInt. Consequently Int32 with value > 0 is required
  instead.

Some explicit Integer, Longint and UInt32 types were replaced by Int32.

Replaced "Byte" type with "UInt8".

Added DWORD cast of Int32 parameter in Windows API WriteFile() call.

TTextPatchWriter.IsPrint method parameter type changed to TCChar from
AnsiChar.

Remove erroneous Cardinal cast of a THandle.

+++ Src/Common/Common.IO.pas
@@ -24,7 +24,7 @@ type
     ///  <summary>Writes binary data from a buffer pointed to by <c>BufPtr</c>
     ///  with length <c>Size</c> bytes. The data is written to the output
     ///  identified by <c>Handle</c>.</summary>
-    class procedure WriteRaw(Handle: THandle; BufPtr: Pointer; Size: Integer);
+    class procedure WriteRaw(Handle: THandle; BufPtr: Pointer; Size: Int32);
     ///  <summary>Writes a Unicode string <c>S</c> to the output identified by
     ///  <c>Handle</c>.</summary>
     class procedure WriteStr(Handle: THandle; const S: UnicodeString); overload;
@@ -76,12 +76,12 @@ begin
 end;
 
 class procedure TCommonIO.WriteRaw(Handle: THandle; BufPtr: Pointer;
-  Size: Integer);
+  Size: Int32);
 begin
   if Size <= 0 then
     Exit;
   var Dummy: DWORD;
-  Winapi.Windows.WriteFile(Handle, BufPtr^, Size, Dummy, nil);
+  Winapi.Windows.WriteFile(Handle, BufPtr^, DWORD(Size), Dummy, nil);
 end;
 
 class procedure TCommonIO.WriteStr(Handle: THandle; const S: UnicodeString);
